Added searchable combobox for all sidebar windows
*Removed some unused code
