+++ src/main/java/UI/SideBarViews/MultiDomainSideBar.java
@@ -6,6 +6,7 @@ import Matrices.Data.Entities.DSMItem;
 import Matrices.Data.Entities.Grouping;
 import Matrices.Data.MultiDomainDSMData;
 import UI.MatrixViews.MultiDomainView;
+import UI.Widgets.DSMItemComboBox;
 import UI.Widgets.Misc;
 import UI.Widgets.NumericTextField;
 import javafx.collections.FXCollections;
@@ -20,6 +21,7 @@ import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.util.Callback;
 import javafx.util.Pair;
+import org.controlsfx.control.SearchableComboBox;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -225,7 +227,7 @@ public class MultiDomainSideBar extends AbstractSideBar {
 
         // set up the items needed by the builder
         HBox connectionsArea = new HBox();
-        ComboBox<DSMItem> itemSelector = new ComboBox<>();
+        DSMItemComboBox itemSelector = new DSMItemComboBox();
         HashMap<CheckBox, DSMItem> connections = new HashMap<>();
         ToggleGroup tg = new ToggleGroup();
         RadioButton selectByRow = new RadioButton("Row");
@@ -235,7 +237,7 @@ public class MultiDomainSideBar extends AbstractSideBar {
         ArrayList<DSMInterfaceType> selectedInterfaces = new ArrayList<>();
 
         // create the viewer for connections
-        WidgetBuilders.createConnectionsViewerScrollPane(
+        createConnectionsViewerScrollPane(
             matrix,
             connectionsArea,
             itemSelector,
@@ -386,7 +388,7 @@ public class MultiDomainSideBar extends AbstractSideBar {
 
         // set up the items needed by the builder
         HBox connectionsArea = new HBox();
-        ComboBox<DSMItem> itemSelector = new ComboBox<>();
+        DSMItemComboBox itemSelector = new DSMItemComboBox();
         HashMap<CheckBox, DSMItem> connections = new HashMap<>();
         ToggleGroup tg = new ToggleGroup();
         RadioButton selectByRow = new RadioButton("Row");
@@ -396,7 +398,7 @@ public class MultiDomainSideBar extends AbstractSideBar {
         ArrayList<DSMInterfaceType> selectedInterfaces = new ArrayList<>();
 
         // create the viewer for connections
-        WidgetBuilders.createConnectionsViewerScrollPane(
+        createConnectionsViewerScrollPane(
             matrix,
             connectionsArea,
             itemSelector,
@@ -593,9 +595,8 @@ public class MultiDomainSideBar extends AbstractSideBar {
         HBox entryArea = new HBox();
 
         // ComboBox to choose which row or column to modify connections of
-        ComboBox<Integer> firstItemSelector = new ComboBox<>();
-        firstItemSelector.setButtonCell(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY.call(null));
-        firstItemSelector.setCellFactory(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY);
+        SearchableComboBox<Integer> firstItemSelector = new SearchableComboBox<>();
+        firstItemSelector.setConverter(DSM_ITEM_COMBOBOX_CONVERTER);
 
         Vector<Integer> items = new Vector<>();
         items.add(Integer.MAX_VALUE);  // this will be used for selecting all items
@@ -616,9 +617,8 @@ public class MultiDomainSideBar extends AbstractSideBar {
         // ComboBox to choose the second item of the connection (row or column)
         ObservableList<Integer> connectionItems = FXCollections.observableArrayList();
 
-        ComboBox<Integer> secondItemSelector = new ComboBox<>();
-        secondItemSelector.setButtonCell(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY.call(null));
-        secondItemSelector.setCellFactory(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY);
+        SearchableComboBox<Integer> secondItemSelector = new SearchableComboBox<>();
+        secondItemSelector.setConverter(DSM_ITEM_COMBOBOX_CONVERTER);
         secondItemSelector.setItems(connectionItems);
 
         secondItemSelector.setMaxWidth(Double.MAX_VALUE);
@@ -627,6 +627,8 @@ public class MultiDomainSideBar extends AbstractSideBar {
 
         // add listener to first item to update options in second column
         firstItemSelector.valueProperty().addListener((observable, oldValue, newValue) -> {
+            if (newValue == null) return;  // ignore null values
+
             if(matrix.isRow(newValue)) {
                 connectionItems.clear();
                 connectionItems.add(Integer.MAX_VALUE);  // add the all option
