+++ src/main/java/UI/SideBarViews/SymmetricSideBar.java
@@ -6,6 +6,7 @@ import Matrices.Data.Entities.DSMItem;
 import Matrices.Data.Entities.Grouping;
 import Matrices.Data.SymmetricDSMData;
 import UI.MatrixViews.SymmetricView;
+import UI.Widgets.DSMItemComboBox;
 import UI.Widgets.Misc;
 import UI.Widgets.NumericTextField;
 import javafx.collections.FXCollections;
@@ -21,11 +22,11 @@ import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
+import org.controlsfx.control.SearchableComboBox;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Vector;
 
 /**
  * Creates a sidebar with methods to interact with a symmetric matrix
@@ -196,7 +197,7 @@ public class SymmetricSideBar extends AbstractSideBar {
 
         // set up the items needed by the builder
         HBox connectionsArea = new HBox();
-        ComboBox<DSMItem> itemSelector = new ComboBox<>();
+        DSMItemComboBox itemSelector = new DSMItemComboBox();
         HashMap<CheckBox, DSMItem> connections = new HashMap<>();
         ToggleGroup tg = new ToggleGroup();
         RadioButton selectByRow = new RadioButton("Row");
@@ -206,7 +207,7 @@ public class SymmetricSideBar extends AbstractSideBar {
         ArrayList<DSMInterfaceType> selectedInterfaces = new ArrayList<>();
 
         // create the viewer for connections
-        WidgetBuilders.createConnectionsViewerScrollPane(
+        createConnectionsViewerScrollPane(
             matrix,
             connectionsArea,
             itemSelector,
@@ -356,7 +357,7 @@ public class SymmetricSideBar extends AbstractSideBar {
 
         // set up the items needed by the builder
         HBox connectionsArea = new HBox();
-        ComboBox<DSMItem> itemSelector = new ComboBox<>();
+        DSMItemComboBox itemSelector = new DSMItemComboBox();
         HashMap<CheckBox, DSMItem> connections = new HashMap<>();
         ToggleGroup tg = new ToggleGroup();
         RadioButton selectByRow = new RadioButton("Row");
@@ -366,7 +367,7 @@ public class SymmetricSideBar extends AbstractSideBar {
         ArrayList<DSMInterfaceType> selectedInterfaces = new ArrayList<>();
 
         // create the viewer for connections
-        WidgetBuilders.createConnectionsViewerScrollPane(
+        createConnectionsViewerScrollPane(
             matrix,
             connectionsArea,
             itemSelector,
@@ -563,11 +564,10 @@ public class SymmetricSideBar extends AbstractSideBar {
         HBox entryArea = new HBox();
 
         // ComboBox to choose which row or column to modify connections of
-        ComboBox<Integer> firstItemSelector = new ComboBox<>();
-        firstItemSelector.setButtonCell(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY.call(null));
-        firstItemSelector.setCellFactory(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY);
+        SearchableComboBox<Integer> firstItemSelector = new SearchableComboBox<>();
+        firstItemSelector.setConverter(DSM_ITEM_COMBOBOX_CONVERTER);
 
-        Vector<Integer> items = new Vector<>();
+        ArrayList<Integer> items = new ArrayList<>();
         items.add(Integer.MAX_VALUE);  // this will be used for selecting all items
         for(DSMItem row : matrix.getRows()) {
             items.add(row.getUid());
@@ -586,9 +586,8 @@ public class SymmetricSideBar extends AbstractSideBar {
         // ComboBox to choose the second item of the connection (row or column)
         ObservableList<Integer> connectionItems = FXCollections.observableArrayList();
 
-        ComboBox<Integer> secondItemSelector = new ComboBox<>();
-        secondItemSelector.setButtonCell(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY.call(null));
-        secondItemSelector.setCellFactory(MATRIX_ITEM_INTEGER_COMBOBOX_CELL_FACTORY);
+        SearchableComboBox<Integer> secondItemSelector = new SearchableComboBox<>();
+        secondItemSelector.setConverter(DSM_ITEM_COMBOBOX_CONVERTER);
         secondItemSelector.setItems(connectionItems);
 
         secondItemSelector.setMaxWidth(Double.MAX_VALUE);
@@ -597,6 +596,8 @@ public class SymmetricSideBar extends AbstractSideBar {
 
         // add listener to first item to update options in second column
         firstItemSelector.valueProperty().addListener((observable, oldValue, newValue) -> {
+            if (newValue == null) return;  // ignore null values
+
             if(matrix.isRow(newValue)) {
                 connectionItems.clear();
                 connectionItems.add(Integer.MAX_VALUE);  // add the all option
